+++ src/utilities/index.ts
@@ -1,4 +1,4 @@
+export * from './BrowserStorage.js';
 export * from './decorators';
 export * from './notifier.js';
-export * from './storage.js';
 export * from './theme.js';