+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
 ## [Unreleased]
-- 
+- No changes yet.
+
+## 0.1.2 (2025-11-13)
+- fixed a type issue in `theme` utility.
+
+## 0.1.1 (2025-11-13)
+- added `BrowserStorage` which supports `localStorage` and `cookie` with async methods for getting, setting, and removing items.
+- changed theme `persist` option to `store` option, and fixed some issues.
+- removed `toggle` method from theme utility.
 
 ## 0.1.0 (2025-11-12)
 - Initial library version release